Refactor spring-ai-client-chat-1.0.0 tests
