+++ instrumentation/spring-ai-client-chat-1.0.0/src/main/java/llm/completions/models/springai/SpringAiModelResponse.java
@@ -187,7 +187,7 @@ public class SpringAiModelResponse implements ModelResponse {
     @Override
     public Integer getTimeToFirstToken() {
         // This only applies to streams
-        return timeToFirstToken; // TODO this is from streams only
+        return timeToFirstToken;
     }
 
     @Override
